+++ Win32ApplicationFramework/lib/source/win32/win32_platform.cpp
@@ -1,7 +1,7 @@
 /*!
 lib\source\win32\win32_platform.cpp
 Created: October 5, 2025
-Updated: November 25, 2025
+Updated: November 29, 2025
 Copyright (c) 2025, Jacob Gosse
 
 Win32 Platform source file.
@@ -23,30 +23,30 @@ class TestWindow : public Window
 {
 public:
     TestWindow(
-        WNDCLASSEX& wcex, 
+        WindowClassRegistry& windowClassRegistry, 
         const std::wstring& title = L"", 
         LONG windowWidth = DEFAULT_SCREEN_WIDTH, 
         LONG windowHeight = DEFAULT_SCREEN_HEIGHT, 
         MessagePumpMode mode = MessagePumpMode::RealTime, 
         int nCmdShow = SW_SHOWDEFAULT
     )
-        : Window(wcex, title, windowWidth, windowHeight, mode, nCmdShow)
+        : Window(windowClassRegistry, title, windowWidth, windowHeight, mode, nCmdShow)
     {
-        std::wcout << L"CONSTRUCTOR: TestWindow(WNDCLASSEX& wcex)\n";
+        std::wcout << L"CONSTRUCTOR: TestWindow(WindowClassRegistry& windowClassRegistry)\n";
     }
 
     TestWindow(
         HINSTANCE hInstance, 
-        WNDCLASSEX& wcex, 
+        WindowClassRegistry& windowClassRegistry,
         const std::wstring& title = L"", 
         LONG windowWidth = DEFAULT_SCREEN_WIDTH, 
         LONG windowHeight = DEFAULT_SCREEN_HEIGHT, 
         MessagePumpMode mode = MessagePumpMode::RealTime, 
         int nCmdShow = SW_SHOWDEFAULT
     )
-        : Window(hInstance, wcex, title, windowWidth, windowHeight, mode, nCmdShow)
+        : Window(hInstance, windowClassRegistry, title, windowWidth, windowHeight, mode, nCmdShow)
     {
-        std::wcout << L"CONSTRUCTOR: TestWindow(WNDCLASSEX& wcex, HINSTANCE hInstance)" << L'\n';
+        std::wcout << L"CONSTRUCTOR: TestWindow(WindowClassRegistry& windowClassRegistry, HINSTANCE hInstance)" << L'\n';
     }
 
     ~TestWindow() noexcept override
@@ -76,34 +76,7 @@ public:
     LRESULT OnDestroy() override { return Window::OnDestroy(); } // parent class OnDestroy() should run AFTER derived class logic
 };
 
-static WNDCLASSEX CreateWindowClass(HINSTANCE hInstance, const std::wstring& className, int extraClassBytes, int extraWindowBytes)
-{
-    WNDCLASSEX wcex = {};
-
-    // window class properties
-    wcex.cbSize = sizeof(WNDCLASSEX);
-    wcex.lpfnWndProc = nullptr;
-    wcex.hInstance = hInstance;
-    wcex.lpszClassName = className.c_str();
-    wcex.lpszMenuName = MAKEINTRESOURCEW(IDR_MAIN_MENU);
-    wcex.style = CS_HREDRAW | CS_VREDRAW | CS_DBLCLKS | CS_OWNDC;
-    wcex.cbClsExtra = extraClassBytes;
-    wcex.cbWndExtra = extraWindowBytes;
-    wcex.hCursor = LoadCursorW(NULL, IDC_ARROW);
-    wcex.hIcon = LoadIconW(hInstance, MAKEINTRESOURCE(IDI_PRIMARY));
-    wcex.hIconSm = LoadIconW(hInstance, MAKEINTRESOURCE(IDI_SMALL));
-    wcex.hbrBackground = (HBRUSH)(COLOR_WINDOW + 1);
-    //wcex.hbrBackground = (HBRUSH)GetStockObject(BLACK_BRUSH);
-    //wcex.hbrBackground = (HBRUSH)GetStockObject(WHITE_BRUSH);
-
-    return wcex;
-}
-
-static void UnregisterWindowClass(const WNDCLASSEX& wcex)
-{
-    THROW_IF_ERROR_CTX(!UnregisterClassW(wcex.lpszClassName, wcex.hInstance), L"Failed to unregister the window class!");
-}
-
+/* wWinMain entry */
 int APIENTRY wWinMain(_In_ HINSTANCE hInstance, _In_opt_ HINSTANCE hPrevInstance, _In_ LPWSTR lpCmdLine, _In_ int nCmdShow)
 {
     UNREFERENCED_PARAMETER(hPrevInstance);
@@ -129,21 +102,21 @@ int APIENTRY wWinMain(_In_ HINSTANCE hInstance, _In_opt_ HINSTANCE hPrevInstance
         // run unit tests
         winxframe::TestRegistry::RunAll();
 
-        // create WNDCLASS object
-        WNDCLASSEX mainWindowClass = CreateWindowClass(hInstance, L"WindowBaseClass", 0, 0);
+        // create the WNDCLASSEX registry
+        WindowClassRegistry windowClassRegistry;
 
         // init window(s)
         LONG windowWidth = 800;
         LONG windowHeight = 600;
-        windows.push_back(std::make_unique<TestWindow>(hInstance, mainWindowClass, L"WINDOW1", windowWidth, windowHeight, MessagePumpMode::RealTime, nCmdShow));
-        //windows.push_back(std::make_unique<TestWindow>(hInstance, mainWindowClass, L"WINDOW2", 600, 400, MessagePumpMode::RealTime, nCmdShow));
-        //windows.push_back(std::make_unique<TestWindow>(hInstance, mainWindowClass, L"WINDOW3", 400, 200, MessagePumpMode::EventDriven, nCmdShow));
+        windows.push_back(std::make_unique<TestWindow>(hInstance, windowClassRegistry, L"WINDOW1", windowWidth, windowHeight, MessagePumpMode::RealTime, nCmdShow));
+        //windows.push_back(std::make_unique<TestWindow>(hInstance, windowClassRegistry, L"WINDOW2", 600, 400, MessagePumpMode::RealTime, nCmdShow));
+        //windows.push_back(std::make_unique<TestWindow>(hInstance, windowClassRegistry, L"WINDOW3", 400, 200, MessagePumpMode::EventDriven, nCmdShow));
 
         /* Main Loop */
         std::wcout << L"Entering the main loop...\n";
-        OutputDebugStringW(L"Entering the main loop...\n");
+        OutputDebugString(L"Entering the main loop...\n");
 
-        timeBeginPeriod(1); // sets system timer resolution to 1 ms
+        timeBeginPeriod(1); // set system timer resolution to 1 ms
 
         while (ManageWindows(windows))
         {
@@ -153,19 +126,20 @@ int APIENTRY wWinMain(_In_ HINSTANCE hInstance, _In_opt_ HINSTANCE hPrevInstance
         timeEndPeriod(1);   // restore system timer resolution
 
         std::wcout << L"Exiting the main loop...\n";
-        OutputDebugStringW(L"Exiting the main loop...\n");
+        OutputDebugString(L"Exiting the main loop...\n");
         /* End of Main Loop */
 
-        // unregister window class
-        UnregisterWindowClass(mainWindowClass);
+        // unregister window classes
+        std::wcout << L"Unregistering window classes.\n";
+        windowClassRegistry.UnregisterAll(hInstance);
     }
     catch (const winxframe::Error& e)
     {
         e.MsgBoxAsync();
         std::wcerr << L"Caught Error (Error class):\n" << e.wwhat() << L'\n';
         std::wcerr << e.LogCauseChain();
-        OutputDebugStringW(e.wwhat());
-        OutputDebugStringW(e.LogCauseChain().c_str());
+        OutputDebugString(e.wwhat());
+        OutputDebugString(e.LogCauseChain().c_str());
 
         if (!windows.empty())
             windows.clear();